Move some components into appropriate folders
